feat: extract ic delete auth
