+++ providers/ic/src/index.ts
@@ -1,3 +1,5 @@
 export type {Components, JSX} from './components';
 
 export * from './providers/auth/auth.ic';
+
+export * from './providers/data/user.ic';
