+++ providers/ic/src/interface.d.ts
@@ -1,3 +1,5 @@
 export * from './components';
 
 export * from './providers/auth/auth.ic';
+
+export * from './providers/data/user.ic';
